Move download method to Api class. Add upload methods #5
