+++ lib/foregit.rb
@@ -6,7 +6,6 @@ module Foregit
   require 'talk_commands'
 
   require 'foreman/api'
-  require 'foreman/download'
 
   SETTINGS = Settings.load_from_file
